Move getsizeof tests to seperate files
