+++ static_frame/test/unit/test_util.py
@@ -2741,89 +2741,5 @@ class TestUnit(TestCase):
         self.assertTrue(is_objectable_dt64(np.array(('0001-01-01',), dtype=DT64_DAY)))
         self.assertTrue(is_objectable_dt64(np.array(('9999-12-31',), dtype=DT64_MS)))
 
-    #---------------------------------------------------------------------------
-    def test_total_getsizeof_int(self) -> None:
-        self.assertEqual(getsizeof_recursive([2]), getsizeof(2))
-
-    def test_total_getsizeof_set(self) -> None:
-        self.assertEqual(
-            getsizeof_recursive([set(['a', 'b', 4])]),
-            sum(getsizeof(e) for e in [
-                'a', 'b', 4,
-                set(['a', 'b', 4])
-            ])
-        )
-
-    def test_total_getsizeof_frozenset(self) -> None:
-        self.assertEqual(
-            getsizeof_recursive([frozenset(['a', 'b', 4])]),
-            sum(getsizeof(e) for e in [
-                'a', 'b', 4,
-                frozenset(['a', 'b', 4])
-            ])
-        )
-
-    def test_total_getsizeof_np_array(self) -> None:
-        self.assertEqual(
-            getsizeof_recursive([np.arange(3)]),
-            getsizeof(np.arange(3))
-        )
-
-    def test_total_getsizeof_frozenautomap(self) -> None:
-        self.assertEqual(
-            getsizeof_recursive([FrozenAutoMap(['a', 'b', 'c'])]),
-            getsizeof(FrozenAutoMap(['a', 'b', 'c']))
-        )
-
-    def test_total_getsizeof_dict(self) -> None:
-        self.assertEqual(
-            getsizeof_recursive([{ 'a': 2, 'b': 3, 'c': (4, 5) }]),
-            sum(getsizeof(e) for e in [
-                'a', 2,
-                'b', 3,
-                'c', 4, 5, (4, 5),
-                { 'a': 2, 'b': 3, 'c': (4, 5) }
-            ])
-        )
-
-    def test_total_getsizeof_tuple(self) -> None:
-        self.assertEqual(getsizeof_recursive([(2, 3, 4)]), sum(getsizeof(e) for e in [
-            2, 3, 4,
-            (2, 3, 4)
-        ]))
-
-
-    def test_total_getsizeof_nested_tuple(self) -> None:
-        self.assertEqual(getsizeof_recursive([(2, 'b', (2, 3))]), sum(getsizeof(e) for e in [
-            2, 'b', 3,
-            (2, 3),
-            (2, 'b', (2, 3))
-        ]))
-
-    def test_total_getsizeof_larger_values_is_larger(self) -> None:
-        a = ('a', 'b', 'c')
-        b = ('abc', 'def', 'ghi')
-        self.assertTrue(getsizeof_recursive([a]) < getsizeof_recursive([b]))
-
-    def test_total_getsizeof_more_values_is_larger_a(self) -> None:
-        a = ('a', 'b', 'c')
-        b = ('a', 'b', 'c', 'd')
-        self.assertTrue(getsizeof_recursive([a]) < getsizeof_recursive([b]))
-
-    def test_total_getsizeof_more_values_is_larger_b(self) -> None:
-        a = ('a', 'b', 'c')
-        b = 'd'
-        self.assertTrue(getsizeof_recursive([a]) < getsizeof_recursive([a, b]))
-
-    def test_total_getsizeof_more_values_is_larger_nested_a(self) -> None:
-        a = ('a', (2, (8, 9), 4), 'c')
-        b = ('a', (2, (8, 9, 10), 4), 'c')
-        self.assertTrue(getsizeof_recursive([a]) < getsizeof_recursive([b]))
-
-    def test_total_getsizeof_more_values_is_larger_nested_b(self) -> None:
-        a = np.array(['a', [2, (8, 9), 4], 'c'], dtype=object)
-        b = np.array(['a', [2, (8, 9, 10), 4], 'c'], dtype=object)
-        self.assertTrue(getsizeof_recursive([a]) < getsizeof_recursive([b]))
-
 if __name__ == '__main__':
     unittest.main()
